requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5

